user story #1347: Implement API for search product by name in all categories or in specified cetegories

+++ src/main/java/com/advantage/online/store/dao/product/ProductRepository.java
@@ -19,7 +19,7 @@ public interface ProductRepository extends DefaultCRUDOperations<Product> {
      * @return entity reference
      */
     @Transactional
-    Product create(String name, String description, int price, String imgUrl, Category category);
+    Product create(String name, String description, double price, String imgUrl, Category category);
     /**
      * Get product categories by categoryId
      *
